Core(Gtk): fix picker not setting initial value
Fix bug when initial value of picker would not be set
because of circular dependency between virtual view and
platform view.

So what happened before was that SetValues changed values
of Gtk widget, that would in turn trigger handler's
OnNativeViewChanged, and it would reset
virtualView.SelectedIndex. And in the end, the initial value
would not be set.
Now virtualView.SelectedIndex is not set when SetValues is
running.

Co-authored-by: Parham <parhaamsaremi@gmail.com>